relicensed the repo to CC0/MIT/Apache triple-licensing (!).
this is primarily to avoid legal difficulties. I will continue
making changes to this repo with the public domain dedication,
but this should not prevent the eventual inclusion of this
source code to the Rust main tree nor the further diverging
(i.e. cannot be ported back to this repo) contribution to
the resulting fork of this source code.